synch with upstream

+++ api/http/handler/endpoints/endpoint_update.go
@@ -22,6 +22,8 @@ type endpointUpdatePayload struct {
 	// URL or IP address where exposed containers will be reachable.\
 	// Defaults to URL if not specified
 	PublicURL *string `example:"docker.mydomain.tld:2375"`
+	// GPUs information
+	Gpus []portainer.Pair
 	// Group identifier
 	GroupID *int `example:"1"`
 	// Require TLS to connect against this environment(endpoint)
@@ -71,20 +73,20 @@ func (payload *endpointUpdatePayload) Validate(r *http.Request) error {
 func (handler *Handler) endpointUpdate(w http.ResponseWriter, r *http.Request) *httperror.HandlerError {
 	endpointID, err := request.RetrieveNumericRouteVariableValue(r, "id")
 	if err != nil {
-		return &httperror.HandlerError{http.StatusBadRequest, "Invalid environment identifier route variable", err}
+		return httperror.BadRequest("Invalid environment identifier route variable", err)
 	}
 
 	var payload endpointUpdatePayload
 	err = request.DecodeAndValidateJSONPayload(r, &payload)
 	if err != nil {
-		return &httperror.HandlerError{http.StatusBadRequest, "Invalid request payload", err}
+		return httperror.BadRequest("Invalid request payload", err)
 	}
 
 	endpoint, err := handler.DataStore.Endpoint().Endpoint(portainer.EndpointID(endpointID))
 	if handler.DataStore.IsErrObjectNotFound(err) {
-		return &httperror.HandlerError{http.StatusNotFound, "Unable to find an environment with the specified identifier inside the database", err}
+		return httperror.NotFound("Unable to find an environment with the specified identifier inside the database", err)
 	} else if err != nil {
-		return &httperror.HandlerError{http.StatusInternalServerError, "Unable to find an environment with the specified identifier inside the database", err}
+		return httperror.InternalServerError("Unable to find an environment with the specified identifier inside the database", err)
 	}
 
 	if payload.Name != nil {
@@ -110,6 +112,10 @@ func (handler *Handler) endpointUpdate(w http.ResponseWriter, r *http.Request) *
 		endpoint.PublicURL = *payload.PublicURL
 	}
 
+	if payload.Gpus != nil {
+		endpoint.Gpus = payload.Gpus
+	}
+
 	if payload.EdgeCheckinInterval != nil {
 		endpoint.EdgeCheckinInterval = *payload.EdgeCheckinInterval
 	}
@@ -133,30 +139,27 @@ func (handler *Handler) endpointUpdate(w http.ResponseWriter, r *http.Request) *
 			removeTags := tag.Difference(endpointTagSet, payloadTagSet)
 
 			for tagID := range removeTags {
-				tag, err := handler.DataStore.Tag().Tag(tagID)
-				if err != nil {
-					return &httperror.HandlerError{http.StatusInternalServerError, "Unable to find a tag inside the database", err}
-				}
+				err = handler.DataStore.Tag().UpdateTagFunc(tagID, func(tag *portainer.Tag) {
+					delete(tag.Endpoints, endpoint.ID)
+				})
 
-				delete(tag.Endpoints, endpoint.ID)
-				err = handler.DataStore.Tag().UpdateTag(tag.ID, tag)
-				if err != nil {
-					return &httperror.HandlerError{http.StatusInternalServerError, "Unable to persist tag changes inside the database", err}
+				if handler.DataStore.IsErrObjectNotFound(err) {
+					return httperror.InternalServerError("Unable to find a tag inside the database", err)
+				} else if err != nil {
+					return httperror.InternalServerError("Unable to persist tag changes inside the database", err)
 				}
 			}
 
 			endpoint.TagIDs = payload.TagIDs
 			for _, tagID := range payload.TagIDs {
-				tag, err := handler.DataStore.Tag().Tag(tagID)
-				if err != nil {
-					return &httperror.HandlerError{http.StatusInternalServerError, "Unable to find a tag inside the database", err}
-				}
+				err = handler.DataStore.Tag().UpdateTagFunc(tagID, func(tag *portainer.Tag) {
+					tag.Endpoints[endpoint.ID] = true
+				})
 
-				tag.Endpoints[endpoint.ID] = true
-
-				err = handler.DataStore.Tag().UpdateTag(tag.ID, tag)
-				if err != nil {
-					return &httperror.HandlerError{http.StatusInternalServerError, "Unable to persist tag changes inside the database", err}
+				if handler.DataStore.IsErrObjectNotFound(err) {
+					return httperror.InternalServerError("Unable to find a tag inside the database", err)
+				} else if err != nil {
+					return httperror.InternalServerError("Unable to persist tag changes inside the database", err)
 				}
 			}
 		}
@@ -187,12 +190,8 @@ func (handler *Handler) endpointUpdate(w http.ResponseWriter, r *http.Request) *
 		switch *payload.Status {
 		case 1:
 			endpoint.Status = portainer.EndpointStatusUp
-			break
 		case 2:
 			endpoint.Status = portainer.EndpointStatusDown
-			break
-		default:
-			break
 		}
 	}
 
@@ -211,7 +210,7 @@ func (handler *Handler) endpointUpdate(w http.ResponseWriter, r *http.Request) *
 		httpClient := client.NewHTTPClient()
 		_, authErr := httpClient.ExecuteAzureAuthenticationRequest(&credentials)
 		if authErr != nil {
-			return &httperror.HandlerError{http.StatusInternalServerError, "Unable to authenticate against Azure", authErr}
+			return httperror.InternalServerError("Unable to authenticate against Azure", authErr)
 		}
 		endpoint.AzureCredentials = credentials
 	}
@@ -255,7 +254,7 @@ func (handler *Handler) endpointUpdate(w http.ResponseWriter, r *http.Request) *
 			endpoint.TLSConfig.TLSKeyPath = ""
 			err = handler.FileService.DeleteTLSFiles(folder)
 			if err != nil {
-				return &httperror.HandlerError{http.StatusInternalServerError, "Unable to remove TLS files from disk", err}
+				return httperror.InternalServerError("Unable to remove TLS files from disk", err)
 			}
 		}
 
@@ -265,11 +264,11 @@ func (handler *Handler) endpointUpdate(w http.ResponseWriter, r *http.Request) *
 		}
 	}
 
-	if payload.URL != nil || payload.TLS != nil || endpoint.Type == portainer.AzureEnvironment {
+	if (payload.URL != nil && *payload.URL != endpoint.URL) || (payload.TLS != nil && endpoint.TLSConfig.TLS != *payload.TLS) || endpoint.Type == portainer.AzureEnvironment {
 		handler.ProxyManager.DeleteEndpointProxy(endpoint.ID)
 		_, err = handler.ProxyManager.CreateAndRegisterEndpointProxy(endpoint)
 		if err != nil {
-			return &httperror.HandlerError{http.StatusInternalServerError, "Unable to register HTTP proxy for the environment", err}
+			return httperror.InternalServerError("Unable to register HTTP proxy for the environment", err)
 		}
 	}
 
@@ -277,35 +276,35 @@ func (handler *Handler) endpointUpdate(w http.ResponseWriter, r *http.Request) *
 		if endpoint.Type == portainer.KubernetesLocalEnvironment || endpoint.Type == portainer.AgentOnKubernetesEnvironment || endpoint.Type == portainer.EdgeAgentOnKubernetesEnvironment {
 			err = handler.AuthorizationService.CleanNAPWithOverridePolicies(endpoint, nil)
 			if err != nil {
-				return &httperror.HandlerError{http.StatusInternalServerError, "Unable to update user authorizations", err}
+				return httperror.InternalServerError("Unable to update user authorizations", err)
 			}
 		}
 	}
 
 	err = handler.DataStore.Endpoint().UpdateEndpoint(endpoint.ID, endpoint)
 	if err != nil {
-		return &httperror.HandlerError{http.StatusInternalServerError, "Unable to persist environment changes inside the database", err}
+		return httperror.InternalServerError("Unable to persist environment changes inside the database", err)
 	}
 
 	if (endpoint.Type == portainer.EdgeAgentOnDockerEnvironment || endpoint.Type == portainer.EdgeAgentOnKubernetesEnvironment) && (groupIDChanged || tagsChanged) {
 		relation, err := handler.DataStore.EndpointRelation().EndpointRelation(endpoint.ID)
 		if err != nil {
-			return &httperror.HandlerError{http.StatusInternalServerError, "Unable to find environment relation inside the database", err}
+			return httperror.InternalServerError("Unable to find environment relation inside the database", err)
 		}
 
 		endpointGroup, err := handler.DataStore.EndpointGroup().EndpointGroup(endpoint.GroupID)
 		if err != nil {
-			return &httperror.HandlerError{http.StatusInternalServerError, "Unable to find environment group inside the database", err}
+			return httperror.InternalServerError("Unable to find environment group inside the database", err)
 		}
 
 		edgeGroups, err := handler.DataStore.EdgeGroup().EdgeGroups()
 		if err != nil {
-			return &httperror.HandlerError{http.StatusInternalServerError, "Unable to retrieve edge groups from the database", err}
+			return httperror.InternalServerError("Unable to retrieve edge groups from the database", err)
 		}
 
 		edgeStacks, err := handler.DataStore.EdgeStack().EdgeStacks()
 		if err != nil {
-			return &httperror.HandlerError{http.StatusInternalServerError, "Unable to retrieve edge stacks from the database", err}
+			return httperror.InternalServerError("Unable to retrieve edge stacks from the database", err)
 		}
 
 		currentEdgeStackSet := map[portainer.EdgeStackID]bool{}
@@ -319,8 +318,13 @@ func (handler *Handler) endpointUpdate(w http.ResponseWriter, r *http.Request) *
 
 		err = handler.DataStore.EndpointRelation().UpdateEndpointRelation(endpoint.ID, relation)
 		if err != nil {
-			return &httperror.HandlerError{http.StatusInternalServerError, "Unable to persist environment relation changes inside the database", err}
+			return httperror.InternalServerError("Unable to persist environment relation changes inside the database", err)
 		}
+	}
+
+	err = handler.SnapshotService.FillSnapshotData(endpoint)
+	if err != nil {
+		return httperror.InternalServerError("Unable to add snapshot data", err)
 	}
 
 	return response.JSON(w, endpoint)
